Ensure apple_itunes_app test only runs for canonical pages
